requirements/development.txt
-r base.txt

# Testing
pytest-flask==0.10.0
pytest-env==0.6.0
